README
This is for the Fastbus portion of the SBS DAQ, which is used by Edet (12 crates) and Cdet (9 crates).  There is a library of fastbus test and diagnostic code from Dave Abbott with small modifications from me.  There is also a readout list which I'm trying to make general purpose -- one readout list for all crates.  The readout list automatically senses what modules there are and what kinds they are.